+++ src/com/kaylerrenslow/armaDialogCreator/arma/stringtable/Language.java
@@ -4,7 +4,6 @@ import org.jetbrains.annotations.NotNull;
 
 /**
  @author Kayler
- @see StringTableValue
  @since 12/12/2016 */
 public interface Language {
 	/** Get name of language */